requirements.txt
Django==2.2.1
redis==3.2.1
djangorestframework==3.9.4